Add a quick sketch of a Releases class; still WIP and not ready for use.
This is more a starting point for discussion. I'm not really sure this ought to be a separate class, but it probably doesn't hurt to separate the code like this.

Since each release of a plugin would be a separate CPT, we can use postmeta to store most of the details.

For now I've made it mostly private, so the CPT shouldn't be visible except via the REST API. Maybe that could change too.

This won't do anything yet, the code isn't called.